[hotfix] Rename and move MesosEntrypointUtils to o.a.f.mesos.util.MesosUtils
